splitted on js in two
